chore(errors): split core.py into core/ subpackage
Extract ErrorMessage, ErrorMetadata, and FieldReference from flat
core.py into individual files under core/ subpackage with re-export
__init__.py. External imports from errors.core continue to work
transparently.
