--- src/forging_blocks/foundation/errors/core.py
@@ -1,27 +1,0 @@
-"""Core error components for the building blocks foundation.
-
-Defines fundamental data structures for error messages, metadata, and field references.
-"""
-
-from dataclasses import dataclass, field
-
-
-@dataclass(frozen=True)
-class ErrorMessage:
-    """Represents an immutable error message component."""
-
-    value: str
-
-
-@dataclass(frozen=True)
-class ErrorMetadata[T]:
-    """Represents metadata about the error."""
-
-    context: dict[str, T] = field(default_factory=lambda: {})
-
-
-@dataclass(frozen=True)
-class FieldReference:
-    """Represents a reference to a field in the error message."""
-
-    value: str
