+++ src/forging_blocks/foundation/errors/core/error_message.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class ErrorMessage:
+    """Represents an immutable error message component."""
+
+    value: str
