+++ src/forging_blocks/foundation/errors/core/field_reference.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class FieldReference:
+    """Represents a reference to a field in the error message."""
+
+    value: str
